### 2022-07-09 #178 @whedon
[PRE REVIEW]: From Maps to Models - Tutorials for structural geological modeling using GemPy and GemGIS
**Submitting author:** @AlexanderJuestel (<a href="http://orcid.org/0000-0003-0980-7479">Alexander Jüstel</a>)
**Repository:** <a href="https://github.com/cgre-aachen/gemgis_data" target ="_blank">https://github.com/cgre-aachen/gemgis_data</a>
**Version:** v.1.0.0
**Managing EiC:** Jordan Gorzalski

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
