### 2022-07-17 #178 @AlexanderJuestel
Dear @jgorzalski,

I would start with **sharondenisse** as her preferred programming language is Python and her Topic Areas include Geology
Secondly, I would go with **jwagemann** as her preferred programming language is also Python and her Topic Areas include Geospatial Data Analysis.

Cheers
Alex
