### 2022-10-24 #178 @labarba
Hi @AlexanderJuestel — I cannot give you a schedule right now. The good news is that JOSE had a campaign to recruit new editors, and we have 7 editors joining the board as we speak! (They are being onboarded this week, and their profiles has already been posted to our website.) Thank you for your patience.
