### 2022-11-22 #178 @acocac
Hi @jwagemann — you volunteered to review for JOSE (The Journal of Open Source Education). Could you take a look at this submission and let us know if you might be able to volunteer a review for us? Thanks! 🙏
Title: "From Maps to Models - Tutorials for structural geological modeling using GemPy and GemGIS"
[View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00178/jose.00178/10.21105.jose.00178.pdf) :page_facing_up: :point_left:
