### 2022-11-25 #178 @acocac
@AlexanderJuestel, thanks for the submission. It looks in line with the scope of JOSE. Note I've already asked to suggested reviewers their availability. I'll try to reach them via e-mail too.
