### 2022-12-01 #185 @whedon
[REVIEW]: From Maps to Models - Tutorials for structural geological modeling using GemPy and GemGIS
**Submitting author:** @AlexanderJuestel (<a href="http://orcid.org/0000-0003-0980-7479">Alexander Jüstel</a>)
**Repository:** <a href="https://github.com/cgre-aachen/gemgis_data" target ="_blank">https://github.com/cgre-aachen/gemgis_data</a>
**Version:** v.1.0.0

- [ ] **Version:** Does the release version given match the repository release (v.1.0.0)?

- [ ] **Version:** Does the release version given match the repository release (v.1.0.0)?
