### 2022-12-19 #185 @BenjMy
@AlexanderJuestel @acocac

I have completed an initial review of the GemGIS tutorial. First I'd like to congratulate the GemGIS dev team for this nice piece of work. I took the measure of all the time and investment made to build the tools and notebooks and make them accessible for everyone.

This package has the potential to make an immediate impact on the scientific communities. There are thoughtful and well-paced tutorial notebooks that provide high-level introductions and more detailed examples of the various methods offered in GemGIS (and Gempy). I am excited both for the package's potential in the long run and possibly to use it in the immediate future.

Below are some of my initial thoughts from the checklist that I will begin creating issues for and linking back to this comment.

# Checklist topics
 - **General Checks**
   - Version: up to now there is no release of the proposed package.
   - Contribution and authorship: unclear what role Miguel de la Varga, Nils Chudalla, and Stefan Back have played in the code development (no commits).

 - **Documentation**

   - Readers are invited to refer to the GEMGIS core code documentation since there is no specific documentation for the GEMGIS-data repos. I think hosting the datasets and notebooks in a separate repo is perfectly relevant particularly to provide clear guidelines on the data **FAIRness**.

     - I suggest adding details (add a section in the documentation) covering how FAIR are the dataset provided:
     - The LICENCE file covers the GemGIS software rights. But under what Licence are the data distributed? I've seen that the authors cited individually the map source for each notebook but it should also clearly be stated on the README and in the doc.
     - Is there a plan for long-term archiving of the data? into Zenodo for instance?
     - Is there a plan for version-controlled data? data could be fetched (using Pooch for example).

   - _A statement of need_
     - The authors don't identify the target audience in the docs nor in the README.
     - The definition of the problem being solved is only part of the doc. Would be also interesting to add it to the README.
   - _Installation instructions_
     - Issues with the badge target link
     - The is no list of dependencies in the README. At least the authors should make clear that Gempy is an optional dependency.
     - Error using conda-forge installation (inconsistent env)
     - Error using pip installation (missing dependencies)
     - In the doc, reorder and put "Installing Gempy (Optional)" at the end of the list.

   - _Usage_
     - I opened an issue with some recommendations.
     - Error Using Binder when GEMGIS is imported.

   - _Community guidelines_:
     - Add "Authorship and credit" statement.
     - I would detail a bit more what can of data can be submitted. Under what terms?

 - **Pedagogy / Instructional design**

   - _Learning objectives_

   Individually each notebook is thoroughly documented to guide the user and it is really much appreciated. Nevertheless, the documentation design with respect to the learning objective can be improved. I recommend the authors articulate the notebooks over sections and subsections. Ideally, I would distinguish within new section tutorials which are Gempy dependent from the others. Also, you could articulate the notebooks using sections following your "Learning Objectives" workflow.

   - _Instructional design_
     - Video Tutorials are expected and I think it is a great initiative.

 - **Jose Paper**

Overall it is a good start but it reads a bit choppy. I found the article very resource exploitation (oil, mining, ...) target-oriented while I think this tool has many other disciplines to reach.

   - _Statement of needs_

     - To me for instance the tools developed can be also associated with environmental risk management (water scarcity, landslides, ...). I am personally interested in this package for Catchment hydrogeology modelling. Authors should broaden this idea in the statement of needs.

     - I like the proof of concept and that the authors already used the software to teach RWTH Aachen. Could it be possible to add (in the "Experience of use in teaching and learning situations" section):
     	- a link to the degree where it has been taught;
     	- the number of hours dedicated to the course (and how it has been distributed over all the notebooks);
     	- the ratio between the number of instructors and the number of students.
     - "Even from seismic data" is to me again to me narrow. Why other geophysical approaches able to infer geological structures (TEM, gravity, ...) are not considered?

   - _Ressources_:
     - I suggest changing the order of resources and placing "GemGIS" first and then adding "GemPy" indicating that it is an optional dependency.
     - "Tutorial repository" link and "GemGIS link" are redundant. Choose one or another or change the first to "https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/index.html"

   - _References_:  it looks like the references were broken during the paper compilation.
