### 2022-12-19 cgre-aachen/gemgis_data#13 @BenjMy
[JOSE] Badges issue
xref https://github.com/openjournals/jose-reviews/issues/185

In the README file, I would expect the badges to link with the proper target i.e.:

- conda forge -> to the conda gemgis forge repo
- pypy -> to the pypy gemgis repo
- (Check also license and doc)
