### 2022-12-19 cgre-aachen/gemgis_data#14 @BenjMy
[JOSE] pypi installation 
xref https://github.com/openjournals/jose-reviews/issues/185

The required dependencies should be installed automatically when you install GemGIS using pip.

`pip install gemgis` runs well but led to missing dependencies after importing GemGIS (matplolib, geopandas,...) .

You would probably need to revise your setup.py to add the necessary requirements.
