### 2022-12-19 cgre-aachen/gemgis_data#15 @BenjMy
[JOSE] add a section in the documentation about FAIRness of the datasets
xref https://github.com/openjournals/jose-reviews/issues/185

Hosting the datasets and notebooks in a separate repo is perfectly relevant particularly to provide clear guidelines on the data FAIRness.

I suggest adding details in the README and in the documentation covering how FAIR are the dataset provided (but also for future contributions), in particular:
- Under what Licence are the data distributed? I've seen that the authors cited individually the map source for each notebook but it should also clearly be stated on the README and in the doc.
- Is there a plan for long-term archiving of the data? into Zenodo for instance?
- Is there a plan for version-controlled data? data could be fetched (using Pooch for example).
