### 2022-12-19 cgre-aachen/gemgis_data#16 @BenjMy
[JOSE] Pyvista Error Using Binder when GEMGIS is imported
xref https://github.com/openjournals/jose-reviews/issues/185

When GemGIS is imported  into the binder it returns the following error:
`ImportError: libGL.so.1: cannot open shared object file: No such file or directory. `

Please check out https://docs.pyvista.org/getting-started/installation.html in order to create the file called **start** for pyvista specific requirements for mybinder.
