### 2022-12-19 cgre-aachen/gemgis_data#17 @BenjMy
[JOSE] First Release and Archive of the datasets
xref https://github.com/openjournals/jose-reviews/issues/185

Yet there is no release of the proposed package and archive of the datasets.
I suggest creating one (even if it is a beta) and linking it to a data repository to assign a DOI (GitHub integration with Zenodo is probably the easiest way see [here](https://docs.github.com/fr/repositories/archiving-a-github-repository/referencing-and-citing-content)
