### 2022-12-19 cgre-aachen/gemgis_data#18 @BenjMy
[JOSE] Pedagogy / Instructional design
xref https://github.com/openjournals/jose-reviews/issues/185

The documentation design with respect to the learning objectives stated in the article can be improved.

I suggest the authors articulate the notebooks over sections and subsections. Ideally, I would distinguish within new section tutorials which are GemPy-dependent from the others. Also, you could articulate the notebooks using sections following your "Learning Objectives" workflow.

Note that individually each notebook is thoroughly documented to guide the user and it is really much appreciated!
