### 2022-12-19 cgre-aachen/gemgis_data#20 @BenjMy
[JOSE] Documentation/Readme possible improvements
xref https://github.com/openjournals/jose-reviews/issues/185

- README
  - Make clear that Gempy is an optional dependency in the README.
  - Identify the target audience in the docs and in the README.
  - Add briefly FAIR principles adopted for the datasets

- Documentation:
  - The installation section is a bit confusing. I would only add 2 clear installations: (1) Installing using Conda/Pip (2) Installation of a developer version. Particularly I don't understand why you recommend "installing Gemgis dependencies manually" after "Installing GemGIS via conda-forge". I would expect the required dependencies installed automatically.
  _Notebooks recommandations_
    - See https://github.com/cgre-aachen/gemgis_data/issues/18
    - I often found very useful to have a button to be able to download the notebook directly from the documentation (see Sphinx doc).
