### 2022-12-19 cgre-aachen/gemgis_data#19 @jwagemann
Hi. I am adding additional comments to each issue @BenjMy already started.

* References in the paper seem not to be properly rendered
* Statement of Need needs refinement and need to be clearer
  * What does ‘paper-based analog approach’ mean?
  * I think the scope and target audience can be a bit broader as I see value in the tutorial being of interest for everyone who is interested in using the two open-source packages GemGIS and GemPy
* Make sure to spell out acronyms - e.g. what are ‘AR-sandboxes’?
* you claim that the material can easily be used by other instructors. However, I doubt that this is the case at the moment. I would suggest to make the notebooks more educational and self-explanatory
* From the paper, it is currently hard to follow and grasp the structure and organisation of the course - this is also due to the fact that the course itself can be restructured. See my comment [here](https://github.com/cgre-aachen/gemgis_data/issues/18)
