### 2022-12-19 cgre-aachen/gemgis_data#18 @jwagemann
* The organisation of the notebooks does not allow an easy understanding what notebooks belong to what unit as described in the paper
* There is a potential to organise the notebooks in a more intuitive way and also to give them a more meaningful nomenclature, instead of e.g. 'example32'
* In general, the notebooks can be made more educational
  * Most of the code is only sub-divided by headings, but most code is not explained what it actually does
  * In my opinion the notebooks do not meet the standard of a well-described, self-explanatory tutorial and need to be reworked in order to be pedagogically meaningful
  * Try to re-organise the content and link previous and next notebooks from within the notebook
  * An index page that provides an overview of notebooks and how they should be followed could be also a good idea
  * Consider also providing an overview of the notebook itself at the beginning of each notebook (with anchor links)
  * Some ideas on how to make notebooks more educational can be found [here](https://www.mdpi.com/2072-4292/14/14/3359)
* You address learners in third person (e.g. ‘in this tutorial the user learnt’) → I suggest to use a direct addressing - ‘you will learn …’
* Each notebooks has an empty code cell at the end —> this needs to be cleaned up
