### 2022-12-19 cgre-aachen/gemgis_data#20 @jwagemann
* The repository misses a description on dependencies and environment settings via e.g. an environment.yml file → how can I replicate the setup to have the notebooks of the tutorial running?
* The link to the binder does not open an environment where I can run the tutorial notebooks → it leads to the GemGIS binder setup
* I cannot see a version indicated on the repository
* License of repository (GPL 3.0) and license identified on the notebooks (CC-BY) differ
