### 2022-12-19 #185 @jwagemann
@acocac I also concluded my review and added my comments to the issues @BenjMy opened in the target repository.

I'd like to thank the authors for submitting their tutorial to JOSE and I think that the notebooks have a great potential to be a valuable contribution for the GemGIS and GemPy communities. However, at this stage, my evaluation is that the tutorials require a major review as in my opinion pedagogical and educational standards of notebooks are not met.

There is room to improve the structure of the course, but also the quality of the notebooks in terms of including a pedagogical narrative. Once the notebooks have been better structured and improved, the paper submission needs a revision as well. There is also potential to improve structure, clarity and guidance.

My proposals on how to improve the submission are found as a response to the following issues:
* https://github.com/cgre-aachen/gemgis_data/issues/20
* https://github.com/cgre-aachen/gemgis_data/issues/19
* https://github.com/cgre-aachen/gemgis_data/issues/18
