### 2022-12-24 cgre-aachen/gemgis_data#14 @AlexanderJuestel
Dear @BenjMy,

thanks for this issue. There is an installation page on how to install GemGIS: https://gemgis.readthedocs.io/en/latest/getting_started/installation.html

I will certainly have to revise this one now that major packages have been updated (e.g. Shapely). We decided on having this step-by-step installation to make sure that every package is properly installed rather than installing everything at once and get problematic errors. GemGIS relies on packages that depend in return on GDAL (geopandas and rasterio) or VTK (pyvista) or Theano (gempy), which can be tricky to install within themselves.

What is your take as a reviewer on that? I would like to keep the installation of GemGIS simple if all major dependencies are properly installed correctly.

Cheers Alex
