### 2022-12-24 cgre-aachen/gemgis_data#13 @AlexanderJuestel
Hello @BenjMy,

thanks for opening this issue. Does your issue imply that you would make the `gemgis_data` repo available on Pypi and via conda forge? As this repo only consists of tutorial materials, we have not considered adding a Pypi and conda forge solution yet.

What is your take on that?
