### 2022-12-24 cgre-aachen/gemgis_data#15 @AlexanderJuestel
Dear @BenjMy,

thanks for opening this issue. This is certainly an important topic.

Let me get straight to your 2rd point. Once the JOSE submission is ready to be accepted, there should be a first archived version which we usually store on Zenodo, So we could update that once major changes or new notebooks have been added.

Regarding your 3rd point: The data is currently hosted within the repository but also on a university OwnCloud storage. The data will then be fetched using pooch actually.

I will work on your first point one of these days then!

Cheers
Alex
