### 2022-12-24 cgre-aachen/gemgis_data#17 @AlexanderJuestel
Hello @BenjMy,

this relates back to #13 where I was wondering if you would like to see a Pypi and conda release. The Zenode integration is certainly something we can implement.

Cheers
Alex
