### 2022-12-28 cgre-aachen/gemgis_data#14 @AlexanderJuestel
Dear @BenjMy,

in combination with your issue here, I will have to refactor some of the code of GemGIS as the pygeos package was merged with Shapely recently. This will make it hopefully a little easier to install GemGIS in the future if one dependency is gone.
