### 2022-12-29 7824529
Create greetings.yml
