+++ .github/workflows/greetings.yml
@@ -1,0 +1,28 @@
+name: Greetings
+
+on:
+  pull_request_target:
+    types:
+      - opened
+  issues:
+    types:
+      - opened
+
+jobs:
+  greeting:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/first-interaction@v1
+        with:
+          repo-token: ${{ secrets.GITHUB_TOKEN }}
+          issue-message: 'Hello and welcome! Thanks for posting your first issue in the GemGIS and GemGIS Data project! Someone from our developers will get back to you. If your question is support related, we may transfer it to the [Discussions](https://github.com/cgre-aachen/gemgis/discussions).'
+          pr-message: >
+            🚀 Thanks for opening your first pull request in GemGIS! 🚀
+            Please make sure you read the [Contributing Guide](https://github.com/cgre-aachen/gemgis/blob/main/CONTRIBUTING.rst) and follow the [Code of Conduct](https://github.com/cgre-aachen/gemgis/blob/main/CODE_OF_CONDUCT.md).
+            A few things to keep in mind:
+              - Remember to run the tests locally to make debugging issues easier.
+              - If you need help writing tests, take a look at the existing ones for inspiration. If you do not know where to start, let us know and we will walk you through it.
+              - All new features should be documented. It helps to write the docstrings for your functions/classes before writing the code. This will help you think about your code design and results in better code.
+              - No matter what, we are really grateful that you put in the effort to open this PR!
+
+# Messages inspired by the PyVista Greeting Messages https://github.com/pyvista/pyvista/blob/main/.github/workflows/greetings.yml
