### 2022-12-30 1b694b1
Create CITATION.cff
