+++ CITATION.cff
@@ -1,0 +1,19 @@
+cff-version: 1.1.0
+message: "If you use this software, please cite it as below."
+authors:
+  - family-names: Jüstel
+    given-names: Alexander
+    orcid: 0000-0003-0980-7479
+  - family-names: de la Varga
+    given-names: Miguel
+    orcid: 0000-0001-6941-2685
+  - family-names: Chudalla
+    given-names: Nils
+  - family-names: Back
+    given-names: Stefan
+  - family-names: Wellmann
+    given-names: Florian
+    orcid: 0000-0003-2552-1876
+title: cgre-aachen/gemgis_data: v0.0.x
+version: v0.0.x
+date-released: 2022-12-30
