### 2022-12-30 cgre-aachen/gemgis_data#17 @AlexanderJuestel
The integration with Zenodo was enabled and I have made a first pre-release. The next pre-release or final release then will contain all your requested changes. I also added a cff file for a better citation: https://github.com/cgre-aachen/gemgis_data/blob/main/CITATION.cff

https://zenodo.org/account/settings/github/repository/cgre-aachen/gemgis_data#

![image](https://user-images.githubusercontent.com/45469915/210047609-791f49e3-ba95-43ed-a9d6-0621961dd193.png)
