+++ CITATION.cff
@@ -1,5 +1,5 @@
-cff-version: 1.1.0
-message: "If you use this software, please cite it as below."
+cff-version: 1.2.0
+message: "If you use the GemGIS Data Repository, please cite it as below."
 authors:
   - family-names: Jüstel
     given-names: Alexander
