+++ SECURITY.md
@@ -1,0 +1,15 @@
+# Security Policy
+
+## Supported Versions
+
+Use this section to tell people about which versions of your project are
+currently being supported with security updates.
+
+| Version | Supported          |
+| ------- | ------------------ |
+| >0.0.1  | :white_check_mark: |
+| <0.0.1  | :x:                |
+
+## Reporting a Vulnerability
+
+If you have found a vulnerability, please oopen an issue at https://github.com/cgre-aachen/gemgis_data/issues.
