+++ requirements.txt
@@ -1,0 +1,5 @@
+# Requirements should be installed via conda-forge
+
+geopandas>=0.12.2
+rasterio>=1.3.4
+pyvista>=0.37.0
