### 2022-12-30 cgre-aachen/gemgis_data#17 @AlexanderJuestel
@BenjMy,

I tried uploading the package to PyPi, however, the repo is too large (see screenshot).

What we could do now is moving all the data to our owncloud and just download the tutorial data using pooch? What would you think about that?

![image](https://user-images.githubusercontent.com/45469915/210092346-3c553cfb-fca5-438e-9253-6f5999779923.png)
