### 2023-01-05 #185 @AlexanderJuestel
Hey @BenjMy,
I would need some feedback on the issues you opened in the GemGIS Data Repo to continue with my revisions ;)

Thanks and Cheers
Alex
