### 2023-01-11 726d155
Update README.md
Prepare PR relative to https://github.com/cgre-aachen/gemgis_data/issues/13
