### 2023-01-11 cgre-aachen/gemgis_data#13 @BenjMy
Hello Alex,
As I'm not sure I made myself clear (sorry for that), I propose this issue be solved after reviewing the opened PR #22 😉
And I agree there is no need to make the gemgis_data repo available on Pypi.
