### 2023-01-11 cgre-aachen/gemgis_data#14 @BenjMy
Hello Alex,

Thanks for the details.

I understand your point and I think everyone experienced with python and pip would understand. But naively if I use `pip install gemgis`, I expect to have all the required dependencies installed... Hope this would be possible with the new version and the shapely integration.
I keep an eye on the Gemgis repo and will do a final check for the next round of the paper review.

Thanks.
