### 2023-01-11 cgre-aachen/gemgis_data#17 @BenjMy
Sounds good to consider pooch to fetch the data.
I don't know how much work is required for that. Actually, I've seen that you already have it implemented on gemgis no?
[gemgis/download_gemgis_data.py](https://github.com/cgre-aachen/gemgis/blob/e99b00138a0226eaeba760568d14eacbc8ed0165/gemgis/download_gemgis_data.py)
