### 2023-01-11 cgre-aachen/gemgis_data#17 @AlexanderJuestel
Hey,

yes, it is already implemented in GemGIS and I have the data already uploaded to our server, so not that much more work is required here, actually.

But since you said, that we do not need a PyPi release of `gemgis_data`, this becomes obsolete and we can keep the data in the repo where all users can see it right away?
