### 2023-01-11 cgre-aachen/gemgis_data#14 @AlexanderJuestel
Hey,

I can certainly try to get all the dependencies installed via pip for the GemGIS repo!
