### 2023-01-11 cgre-aachen/gemgis_data#13 @AlexanderJuestel
Yeah, somehow, one was created but since there is no code in the repo and only data the previous question came up. But then I will continue working without having in mind to publish the gemgis_data repo on PyPi and fix the installation of gemgis with pip in the other issue and continue then with revising the remaining open tasks for me.

Thanks for the clarification
