### 2023-01-14 #185 @AlexanderJuestel
I copied your @BenjMy  list to better track my progress.

> # Checklist topics
> * **General Checks**
>
>

> - [x] Version: up to now there is no release of the proposed package. --> **As agreed in https://github.com/cgre-aachen/gemgis_data/issues/17, there will be no further release of the package as the package does not contain any code but only data**
> - [x] Contribution and authorship: unclear what role Miguel de la Varga, Nils Chudalla, and Stefan Back have played in the code development (no commits). --> **A author statement was added to the JOSE Paper , see https://github.com/cgre-aachen/gemgis_data/commit/32e6fcafb79d17f9801aa5f07730be4f65e3930a**

> * **Documentation**
>
>   * Readers are invited to refer to the GEMGIS core code documentation since there is no specific documentation for the GEMGIS-data repos. I think hosting the datasets and notebooks in a separate repo is perfectly relevant particularly to provide clear guidelines on the data **FAIRness**.
>
> - [ ]  I suggest adding details (add a section in the documentation) covering how FAIR are the dataset provided:
> - [ ]  The LICENCE file covers the GemGIS software rights. But under what Licence are the data distributed? I've seen that the authors cited individually the map source for each notebook but it should also clearly be stated on the README and in the doc.
> - [x]  Is there a plan for long-term archiving of the data? into Zenodo for instance? --> **A Zenodo Repository will be created upon acceptance of the paper**
> - [x]  Is there a plan for version-controlled data? data could be fetched (using Pooch for example). --> **All files are already stored on a OwnCloud server and accessible via pooch, the notebooks will be adapted to either use the shipped data in the repository or allow the user to download the data for each notebook**
>   * _A statement of need_
>
> - [ ]  The authors don't identify the target audience in the docs nor in the README.
> - [ ]  The definition of the problem being solved is only part of the doc. Would be also interesting to add it to the README.
>   * _Installation instructions_
>
> - [ ] Issues with the badge target link
> - [ ]  The is no list of dependencies in the README. At least the authors should make clear that Gempy is an optional dependency.
> - [ ]  Error using conda-forge installation (inconsistent env)
> - [ ]  Error using pip installation (missing dependencies)
> - [ ]  In the doc, reorder and put "Installing Gempy (Optional)" at the end of the list.
>   * _Usage_
>
> - [ ]  I opened an issue with some recommendations.
> - [ ]  Error Using Binder when GEMGIS is imported.
>   * _Community guidelines_:
>
> - [x]  Add "Authorship and credit" statement. -->  **A author statement was added to the JOSE Paper , see https://github.com/cgre-aachen/gemgis_data/commit/32e6fcafb79d17f9801aa5f07730be4f65e3930a**

> - [ ]  I would detail a bit more what can of data can be submitted. Under what terms?
> * **Pedagogy / Instructional design**
>
>   * _Learning objectives_
>
>   Individually each notebook is thoroughly documented to guide the user and it is really much appreciated. Nevertheless, the documentation design with respect to the learning objective can be improved. I recommend the authors articulate the notebooks over sections and subsections. Ideally, I would distinguish within new section tutorials which are Gempy dependent from the others. Also, you could articulate the notebooks using sections following your "Learning Objectives" workflow.
>
>   * _Instructional design_
>
>     * Video Tutorials are expected and I think it is a great initiative.
> * **Jose Paper**
>
> Overall it is a good start but it reads a bit choppy. I found the article very resource exploitation (oil, mining, ...) target-oriented while I think this tool has many other disciplines to reach.
>
> * _Statement of needs_
>
> - [ ] To me for instance the tools developed can be also associated with environmental risk management (water scarcity, landslides, ...). I am personally interested in this package for Catchment hydrogeology modelling. Authors should broaden this idea in the statement of needs.
> - [ ]  I like the proof of concept and that the authors already used the software to teach RWTH Aachen. Could it be possible to add (in the "Experience of use in teaching and learning situations" section):
>
> - [ ]  a link to the degree where it has been taught;
> - [ ]  the number of hours dedicated to the course (and how it has been distributed over all the notebooks);
> - [ ]  the ratio between the number of instructors and the number of students.
> - [ ]  "Even from seismic data" is to me again to me narrow. Why other geophysical approaches able to infer geological structures (TEM, gravity, ...) are not considered?
> * _Ressources_:
>
> - [x]  I suggest changing the order of resources and placing "GemGIS" first and then adding "GemPy" indicating that it is an optional dependency. --> **The order was changed, see https://github.com/cgre-aachen/gemgis_data/commit/32e6fcafb79d17f9801aa5f07730be4f65e3930a**
> - [x]  "Tutorial repository" link and "GemGIS link" are redundant. Choose one or another or change the first to "https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/index.html" --> **The links were combined, see https://github.com/cgre-aachen/gemgis_data/commit/32e6fcafb79d17f9801aa5f07730be4f65e3930a**
> - [x]  _References_:  it looks like the references were broken during the paper compilation. --> **The references compile perfectly using Github Actions in the main repository, see https://github.com/cgre-aachen/gemgis_data/actions/runs/3918599262**
