### 2023-01-15 cgre-aachen/gemgis_data#23 @AlexanderJuestel
[JOSE] Applications of structural models created with GemPy
xref https://github.com/openjournals/jose-reviews/issues/185

@BenjMy
Copying this over from your comment.  What you have mentioned below are specific applications or how the structures created with GemPy can be further used. Hydrogeological questions (catchment area) cannot be answered with GemPy/GemGIS directly. It is just made for creating the structures of the subsurface. The rocks in between can then be parametrized for the different purposes the structural model is needed for (hydrogeological questions, geothermal, hydrocarbons etc.). **However, with your comment on landslides, I would totally agree to add the broader topic of assessing geohazards to the statement of need.** This could go as far as modeling the structures of volcanoes, or the seafloor to estimate the height of a tsunami or whatever. However, the actual modeling of these questions cannot be done in GemPy/GemGIS or is at least not yet implemented ;)  Could we agree on that? :)

> Overall it is a good start but it reads a bit choppy. I found the article very resource exploitation (oil, mining, ...) target-oriented while I think this tool has many other disciplines to reach. --> **I partially disagree with the statement, that the article is focused on hydrocarbons and mining (professionally, I am not even involved in the HC industry...). Here, the first paragraph of the Statement of need which also includes hydrogeological applications (extraction of fresh water), geothermal applications (extraction of thermal waters), and storage applications (storage of CO2 or hydrogen or nuclear waste):**

_The subsurface below our feet is utilized in many different ways. We extract fresh water, thermal waters and fossil fuels from it. We exploit the subsurface for its coal, minerals and ores in open-pit mines and underground mines. We store fluids and gases as well as nuclear waste in the subsurface. But to do so, a concept of the structure of the subsurface and its properties is needed to drill into the right formation, to dig or drill into the right direction, to ensure that none of the stored fluid escapes and to ensure that no contaminants leak into surrounding rocks._
**
>
> * _Statement of needs_
>
> - [ ] To me for instance the tools developed can be also associated with environmental risk management (water scarcity, landslides, ...). I am personally interested in this package for Catchment hydrogeology modelling. Authors should broaden this idea in the statement of needs.
