### 2023-01-15 cgre-aachen/gemgis_data#20 @AlexanderJuestel
Also, here a list of your comments to track my progress of the revision:

- [x] Make clear that Gempy is an optional dependency in the README.
- [x] Identify the target audience in the docs and in the README.
- [ ] Add briefly FAIR principles adopted for the datasets

- [ ] The installation section is a bit confusing. --> Fixing the installation

- [ ] I often found very useful to have a button to be able to download the notebook directly from the documentation (see Sphinx doc).

- [x] The repository misses a description on dependencies
- [ ] The repository misses a description environment settings via e.g. an environment.yml file → how can I replicate the setup to have the notebooks of the tutorial running?

- [ ] The link to the binder does not open an environment where I can run the tutorial notebooks → it leads to the GemGIS binder setup

- [ ] I cannot see a version indicated on the repository

- [ ] License of repository (GPL 3.0) and license identified on the notebooks (CC-BY) differ
