### 2023-01-15 cgre-aachen/gemgis_data#19 @AlexanderJuestel
But here also a list to track my progress:

   - _Statement of needs_

- [ ] To me for instance the tools developed can be also associated with environmental risks management (water scarcity, landslides, ...). I am personnaly interested to this package for Catchement hydrologeology modelling. Authors should broaden this idea in the statement of needs. --> **This also relates to #23**

- [ ] I like the proof of concept and that authors already used the software for teaching to RWTH Aachen. Could it be possible to add (in the "Experience of use in teaching and learning situations" section):

     	- a link to the degree where it has been tought;
     	- the number of hours dedicated to the course (and how it has been distributed over all the notebooks);
     	- the ratio between the number of instructors and number of students.

- [ ]  "Even from seismic data" is to me again to me to narrow. Why other geophysical approaches able to infer geological structures (TEM, gravity, ...) are not considered?

   - _Ressources_:

- [x] I suggest to change the order of ressources and place "GemGIS" first and then add "GemPy" indicating that it is an optionnal dependency.

- [x]  "Tutorial repository" link and "GemGIS link" are redundant. Choose one or another or change the first to "https://gemgis.readthedocs.io/en/latest/getting_started/tutorial/index.html"

   - _References_:

- [ ]  Seems that references are broken in the paper compilation. --> **The references compile using Github Actions in the gemgis_data repo**
