### 2023-01-24 #185 @AlexanderJuestel
Dear @acocac, dear @BenjMy,

thanks for checking in! I am making progress and the comments are very good! As you can see from the link above (https://github.com/openjournals/jose-reviews/issues/185#issuecomment-1382733498), I am still working on some points here that may take some more or less time. I will try to work on them in a timely manner but my PhD is keeping my quite busy these days ;)

I also have to rework the notebooks still but maybe we can check off some open points once I am done with the none-notebook related issues, also to see what is left :)

Cheers Alex
