### 2023-02-07 cgre-aachen/gemgis_data#15 @AlexanderJuestel
A section about the FAIR principles was added to the Documentation in https://github.com/cgre-aachen/gemgis/commit/be5d3a2f45475a46a2b96f8d08f7dccaad46dfa5

Zenodo archiving is set up already. Pooch is also used to fetch the data. The license issue is discussed these days.
