### 2023-02-07 #185 @AlexanderJuestel
Dear @acocac,

I would like to ask @jwagemann and @BenjMy to go through their comments and remarks and check what I have already completed. In addition, I am sure that some of the tasks at the top of the issue can be checked off already.

It would be amazing to then have an updated list of what I still have to tackle as things are going all over the place by now

Cheers
Alex
