### 2023-02-08 cgre-aachen/gemgis_data#14 @AlexanderJuestel
Hello @BenjMy,

after some consideration, I think we should **not** try to install GemGIS entirely via pip. If you look at the installation instructions for geopandas for using pip, no especially new user would want to "compile and install th[e] C dependencies [of the geopandas dependencies] manually". If you have the most important dependencies installed via conda (geopandas, pyvista, rasterio), then the installation of gemgis via pip is a charm. And we have the installation instructions for gemgis which are easy to follow.

![image](https://user-images.githubusercontent.com/45469915/217458546-8b610f6a-7a2f-448f-ad8f-f448a4aa3545.png)
