### 2023-02-08 #185 @acocac
Hi @jwagemann and @BenjMy. Following @AlexanderJuestel update, I kindly invite you to revisit the checklists and confirm if all suggestions have been addressed in the resource and/or paper.
