### 2023-02-17 cgre-aachen/gemgis_data#23 @BenjMy
Dear @AlexanderJuestel ,

Thanks for the explanation. I see that you made some minor changes in the paper to integrate this comment but also in the readme. Thanks for that. I also found the "Purpose of structural geological models in Earth Sciences" paragraph of the readme really interesting, well done!

I will read the article entirely to check how all the changes integrate but to me, this issue can be closed.
