### 2023-02-17 cgre-aachen/gemgis_data#19 @AlexanderJuestel
@BenjMy, thanks for that! I think I forgot to go over @jwagemann comments here. I will add them to my list above and will go through them right away
