### 2023-02-17 cgre-aachen/gemgis_data#14 @BenjMy
Hi @AlexanderJuestel ,

Thanks for your effort in trying to dig out some possible solutions. I understand your concern about a full pip installation. As soon as the instructions and steps are super clear then I don't think it is an issue anymore (feel free to close).
Thanks.
