### 2023-02-17 #185 @BenjMy
Hi again @acocac  and @AlexanderJuestel ,

I went through the items and validated almost everything except "Pedagogy" and "Instructional Design" where still some efforts are necessary. I guess it is in a good way, I've seen a lot of improvement already for the documentation and readme.

I'll be happy to read again the paper entirely and check the flow of the documentation/notebooks once https://github.com/cgre-aachen/gemgis_data/issues/18 will be tackled.

Well done!
