### 2023-02-17 #185 @acocac
> Hi again @acocac and @AlexanderJuestel ,
>
> I went through the items and validated almost everything except "Pedagogy" and "Instructional Design" where still some efforts are necessary. I guess it is in a good way, I've seen a lot of improvement already for the documentation and readme.
>
> I'll be happy to read again the paper entirely and check the flow of the documentation/notebooks once [cgre-aachen/gemgis_data#18](https://github.com/cgre-aachen/gemgis_data/issues/18) will be tackled.
>
> Well done!

Thanks @BenjMy for all your efforts. @AlexanderJuestel please notify both reviewers when you address the suggested issue.
