### 2023-02-18 cgre-aachen/gemgis_data#14 @AlexanderJuestel
The instructions were reworked and should be clear now, closing the issue
