### 2023-02-18 cgre-aachen/gemgis_data#15 @AlexanderJuestel
Closing this issue as a FAIR section was added
