### 2023-02-25 #185 @AlexanderJuestel
I am making good progress on reworking the tutorial notebooks. I hope to be done with it in one to two weeks if my work schedule allows for that!

![image](https://user-images.githubusercontent.com/45469915/221350732-a623e188-44fa-48e9-b586-d44f25508918.png)
![image](https://user-images.githubusercontent.com/45469915/221350756-487849f1-a929-4a91-97b6-ff25e4dbd846.png)
