### 2023-03-05 #185 @AlexanderJuestel
@acocac, @BenjMy, @jwagemann I have just completed the reworking of the notebooks necessary for the tutorial series. I will revise the manuscript a little but everything is ready for you to review it again I would say. I hope I could meet your expectations but are happy to add some more content if needed

Cheers
Alex
