### 2023-03-06 cgre-aachen/gemgis_data#19 @AlexanderJuestel
@jwagemann I tried to add some information to the manuscript based on your comments. Please feel free to review the changes and let me know if I can close this issue here
