### 2023-03-06 cgre-aachen/gemgis_data#18 @AlexanderJuestel
@jwagemann and @BenjMy, I have worked on your comments and would like to ask you to have a look at the reworked notebooks.
