### 2023-03-06 #185 @acocac
> @acocac, @BenjMy, @jwagemann I have just completed the reworking of the notebooks necessary for the tutorial series. I will revise the manuscript a little but everything is ready for you to review it again I would say. I hope I could meet your expectations but are happy to add some more content if needed
>
> Cheers Alex

Thanks for the update @AlexanderJuestel. I kindly ask @BenjMy @jwagemann to revisit their checklists.
