### 2023-03-31 #185 @BenjMy
Dear @AlexanderJuestel and Gemgis team,

Congrats again on your hard work and dedication in creating this valuable package. I am certain that it will be enlightening for many users. However, I do have a few minor issues and comments that I would like to bring to your attention.

- Could you consider improving "_The organization of the notebooks does not allow an easy understanding of what notebooks belong to what unit as described in the paper_".
I see you renamed the notebooks but could you somehow guide the reader with sections and subsections? This join https://github.com/cgre-aachen/gemgis_data/issues/18 of @jwagemann's issue that for me is not yet completely tackled.
This could be (just as an indicative example but up to you to appreciate better how to name the section):
  - Data extraction methods
  - Meshing tools
  - Processing from Pyvista

- Add the reference to the data source for each notebooks in "tutorial and basics usage" (just like you did for the "example model" section)
- Notebooks 59 to 61 are empty
- Missing logo on the documentation page

- in the documentation, section installation: I would rephrase "Installing GemGIS and its dependencies manually" by "Installing GemGIS's dependencies manually" as GEMGIS has already been installed the line before

- api_reference section is empty in the documentation

Let me know if you want me to create issues on the repo directly.
Best,

Benjamin
