### 2023-03-31 #185 @AlexanderJuestel
@BenjMy thanks for your comments! No need to create issues I think :)

I will fix the documentation, it is actually broken currently due to reasons I still have to figure out....
- Logo: the link is provided in the conf.py file and the file is also provided but I have no idea why it is not showing up. The favicon does though
- The latest notebooks have been added to the main repository and will show up once the documentation compiles again.
- I will fix the installation documentation
- I tried adding an API reference but failed so far. I had one once before but sphinx is playing some games here
- I can add the references to the tutorial notebooks even though they are actually not a part of the course but good hint :)

Regarding the organization of the Notebooks, I added a table of contents in the introduction notebook and added folders for each section mentioned in the introduction notebook. Within several of the notebooks, I tried to include some post-processing tasks rather than having a dedicated post-processing section.

![image](https://user-images.githubusercontent.com/45469915/229116967-11a91c09-5c92-4e32-bcc6-2016ee05c2f6.png)
![image](https://user-images.githubusercontent.com/45469915/229117132-3070bddf-eab6-4809-b44c-8d6f291fa7cb.png)
