### 2023-04-13 #185 @AlexanderJuestel
I have opened an issue with respect to the failing compilation of the documentation here: https://github.com/executablebooks/sphinx-book-theme/issues/719

Waiting on a fix now :)
