### 2023-04-19 #185 @acocac
> > I have opened an issue with respect to the failing compilation of the documentation here: [executablebooks/sphinx-book-theme#719](https://github.com/executablebooks/sphinx-book-theme/issues/719)
> > Waiting on a fix now :)
>
> The docs are working again

@BenjMy may I ask you to revisit your review checklist and comments according to the latest update from @AlexanderJuestel?
