### 2023-05-03 #185 @BenjMy
Hello @acocac and @AlexanderJuestel ,

Thanks to the GemGIS team for their wonderful work. As you can see I ticked all my checklist and all my issues were tackled. I'm already a happy user of GemGIS and hope for the best for its future dev. I have no further comment and I'm looking forward to seeing soon the JOSE paper online.
