### 2023-05-06 #185 @jwagemann
Hi @AlexanderJuestel ,
thank you for the great revisions to the tutorial. It is now much clearer and also easier to follow for a reader / learner.

Before I can tick all boxes, I have some small suggestions for improvement:
* Version: I still see a version mismatch - last release of the repo was on 29 December 2022 with v.0.0.1?
* README needs to be updated:
  * Target audience: there is a spelling mistake ‘Lectures’ and it should read ‘Lecturers’
  * List of dependencies on README:
       * Would be good to list the packages together with their versions used
  * The outline seems to be slightly different from the outline [here](https://github.com/cgre-aachen/gemgis_data/blob/main/notebooks/00_introduction_to_structural_modeling.ipynb), correct?
* Links on the outline needs to be updated:
  * Try to make it ‘browsable’ via nbviewer: https://nbviewer.org/github/cgre-aachen/gemgis_data/blob/main/notebooks/00_introduction_to_structural_modeling.ipynb#installation-of-gempy-and-gemgis
  * You have to use relative paths for the notebook links: If I want to open a notebook under 2 or 3, I get referred back to Github

I did not revise the paper, as it seems that it has not been updated, correct? The only version I found was from 1 December 2022.
