+++ requirements.txt
@@ -1,5 +1,5 @@
 # Requirements should be installed via conda-forge
 
-geopandas>=0.12.2
-rasterio>=1.3.4
-pyvista>=0.37.0
+geopandas>=0.13
+rasterio>=1.3.6
+pyvista>=0.39.0
