### 2023-05-06 #185 @AlexanderJuestel
@jwagemann thanks for checking some boxes! Let me get back to your comments right away!

- I would suggest a bumping of the version and a new release once the paper is ready for publication. Would that be okay with you?

- The paper was updated a couple of times. I have attached the latest build of the paper from 3 weeks ago (https://github.com/cgre-aachen/gemgis_data/actions/runs/4689904821)

- The spelling mistake was fixed in https://github.com/cgre-aachen/gemgis_data/commit/0256472d714cc8feff01da62a74b4add4890d21e

- The main dependencies and their tested versions were added in https://github.com/cgre-aachen/gemgis_data/commit/5b635a267102045174fb2ab788fa7205e66c84e0

- I adapted the outline and added NBViewer links to the notebooks linked in the readme in https://github.com/cgre-aachen/gemgis_data/commit/cc91fd46b19c9972937f15553114b82947555547

I will work on the relative paths tomorrow and post an update as soon as possible :)

[paper.zip](https://github.com/openjournals/jose-reviews/files/11413063/paper.zip)
