### 2023-05-06 #185 @jwagemann
Hi @AlexanderJuestel ,
thanks for you swift response. I think you have to activate the command `@whedon generate pdf` to regenerate your pdf in the JOSE format. I had a look at the md file and I ticked the remaining boxes.

@acocac All good on my side, after Alex changed the paths of the notebooks. I can't tick the box with the version, but I am happy as Alex proposed, to change the version to 1.0.0 once the paper will be published.
