### 2023-05-07 #185 @AlexanderJuestel
Good morning @acocac and @jwagemann,

I changed the links according to the recommendations of @jwagemann in https://github.com/cgre-aachen/gemgis_data/commit/2c4666f2b68bbaa80cba0f75795cc693d3ff9582. To be honest, I did not know NBViewer before and this adds a lot of value to everything! Thanks for the hint!

If you are fine with everything now, I would create a 1.0.0 release and the submission would be ready for publication from my side :)
