### 2023-05-08 #185 @acocac
> If you are fine with everything now, I would create a 1.0.0 release and the submission would be ready for publication from my side :)

@jwagemann, only a last item in your checklist, can you confirm if everything is ok with the version? thank you
