+++ CITATION.cff
@@ -1,21 +1,22 @@
 cff-version: 1.2.0
 message: "If you use the GemGIS Data Repository, please cite it as below."
 authors:
-  - family-names: Jüstel
-    given-names: Alexander
+  - family-names: "Jüstel"
+    given-names: "Alexander"
     orcid: 0000-0003-0980-7479
-  - family-names: de la Varga
-    given-names: Miguel
+  - family-names: "de la Varga"
+    given-names: "Miguel"
     orcid: 0000-0001-6941-2685
-  - family-names: Chudalla
-    given-names: Nils
-  - family-names: Wagner
-    given-names: Jan David
-  - family-names: Back
-    given-names: Stefan
-  - family-names: Wellmann
-    given-names: Florian
+  - family-names: "Chudalla"
+    given-names: "Nils"
+  - family-names: "Wagner"
+    given-names: "Jan David"
+  - family-names: "Back"
+    given-names: "Stefan"
+  - family-names: "Wellmann"
+    given-names: "Florian"
     orcid: 0000-0003-2552-1876
 title: cgre-aachen/gemgis_data: v1.0.0
 version: v1.0.0
 date-released: 2023-05-09
+url: "https://github.com/cgre-aachen/gemgis_data"
