+++ CITATION.cff
@@ -20,3 +20,4 @@ title: "cgre-aachen/gemgis_data: 1.0.0"
 version: 1.0.0
 date-released: 2023-05-09
 url: "https://github.com/cgre-aachen/gemgis_data"
+doi: 10.5281/zenodo.7494025
