### 2023-05-09 #185 @AlexanderJuestel
I have created a 1.0.0 release: https://github.com/cgre-aachen/gemgis_data/releases/tag/v1.0.0

The Zenodo integration failed right now. I am working on a fix!

![image](https://user-images.githubusercontent.com/45469915/237060289-f365cdbb-45fa-4a14-bf8b-0dd97ada2969.png)
