### 2023-05-12 #185 @AlexanderJuestel
@acocac, do you think we can proceed with the publication of the article these days?

Cheers
Alex
