### 2023-05-12 #185 @acocac
> @acocac, do you think we can proceed with the publication of the article these days?
>
> Cheers Alex

Hi @AlexanderJuestel, I'm trying to follow the editorial procedure of JOSE. Please @jwagemann confirm all is ok with the version and complete the checklist 🙏
