### 2023-08-08 #185 @labarba
Hi everybody! 👋
Thank you all for your patience. I looked over the paper, and see that the citations need fixing.
@AlexanderJuestel – Please have a look a the [Markdown citation format](https://pandoc.org/MANUAL.html#citations) we use.
