### 2023-08-08 #185 @AlexanderJuestel
@labarba I have fixed the citation commands in the manuscript and also adapted the rules for capitalization in titles in the bibtex file.

Let me know if there is anything else.

Cheers Alex
