### 2023-08-08 #185 @labarba
We have another error. Please check the report above. It says:

`Error reading bibliography file paper.bib:
(line 175, column 1):
unexpected '@'`
