### 2023-08-09 #185 @labarba
This is such an impressive piece of work!

I could publish as is, but maybe you want to correct a couple of remaining details: on p.3, line 119, the phrase "for the basic structural geological," seems to be missing a noun that follows, like "model" or "map" or something like that - on p.6, line 165, "the open-source developed 3D creation suite Blender" sounds awkward; maybe just delete "developed" - in the Bibliography, use brackets to protect the author display that now awkwardly shows "Community, B. O"
