### 2023-08-15 #185 @labarba
Congratulations @AlexanderJuestel, your JOSE paper is published! 🚀

Huge thanks to our Editor: @acocac and Reviewers: @BenjMy, @jwagemann — your contributions make this possible 🙏
